+++ src/main/java/tc/lv/dao/implementations/IpAddressDaoImpl.java
@@ -40,10 +40,10 @@ public class IpAddressDaoImpl extends DaoAbstract implements IpAddressDao {
     }
 
     @Override
-    public Integer countBlackListByCountyName(String contryName, IpQueryEnum myType) {
+    public Long countBlackListByCountyName(String contryName, IpQueryEnum myType) {
         Query query = entityManager.createNamedQuery(myType.countWhiteOrBlackListByCountry());
         query = query.setParameter(1, false).setParameter(2, contryName);
-        return (Integer) query.getSingleResult();
+        return (Long) query.getSingleResult();
     }
 
     @Override
@@ -53,10 +53,10 @@ public class IpAddressDaoImpl extends DaoAbstract implements IpAddressDao {
     }
 
     @Override
-    public Integer countWhiteListByCountyName(String contryName, IpQueryEnum myType) {
+    public Long countWhiteListByCountyName(String contryName, IpQueryEnum myType) {
         Query query = entityManager.createNamedQuery(myType.countWhiteOrBlackListByCountry());
         query = query.setParameter(1, true).setParameter(2, contryName);
-        return (Integer) query.getSingleResult();
+        return (Long) query.getSingleResult();
     }
 
     @Override
